requirements.txt
dnspython>=2.4.0
httpx[http2]>=0.25.0
requests>=2.31.0
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0
pytest>=7.4.0
aiohttp>=3.9.0
h2>=4.1.0
cryptography>=41.0.0
